ai_service/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.0
transformers==4.44.0
torch==2.4.0
pydantic==2.8.0
grpcio==1.65.0
grpcio-tools==1.65.0
openai==1.40.0
